second commit

+++ app/src/main/java/com/example/simsapp/model/Inventory.kt
@@ -1,7 +1,11 @@
 package com.example.simsapp.model
 
+import com.github.kittinunf.fuel.core.ResponseDeserializable
+import com.google.gson.Gson
+
 data class Inventory(
     val id:Int,
     val description:String,
     val task:Int,
-    val version:Double)
+    val version:Int)
+
